[修改] 調整 Handle 架構
